Introduce unique user tokens using UUID
Using the UUID package from hex, this commit introduces a unique token
for each user which is stored in the session.

A signed token is generated and output to the browser as a meta tag in
the page, which is then used when connecting to the socket with
javascript. This enables us to establish some very basic security and
provides a unique user identifier for each session.