Rock Paper Scissors
Implemented very crude action version of Rock Paper Scissors with a
Triangle, Square and Circle. This version builds on the supervision
tree branch and implements player controls, automatic spawning of
sprites from initial variables, and spawning new sprites when a player
joins, and destroying a sprite when a player leaves, either voluarily by
clicking or by being destroyed by another player. The choice of which
character you are is randomly selected, and the rules follow this:

Triangle -> Circle -> Square -> Triangle